ran python script

+++ app/format_usb_storage.py
@@ -5,6 +5,8 @@ import sys
 import subprocess
 import datetime
 
+DRY_RUN = True
+
 
 def main():
     devname = os.environ.get('DEVNAME')
@@ -16,9 +18,15 @@ def main():
     if not label:
         label = 'USBDRIVE_' + datetime.date.today().strftime('%Y%m%d')
 
-    subprocess.check_output(
-        ['sudo', 'mkfs.fat', '-F', '32', devname, '-n', label]
-    )
+    command = ['sudo', 'mkfs.fat', '-F', '32', devname, '-n', label]
+    print('Command:', command)
+
+    if not DRY_RUN:
+        print('Formatting...')
+        output = subprocess.check_output(
+            ['sudo', 'mkfs.fat', '-F', '32', devname, '-n', label]
+        )
+        print('Output:', output)
 
 
 if __name__ == '__main__':
